+++ framework/include/postprocessors/ChangeOverTimestepPostprocessor.h
@@ -15,7 +15,7 @@
 #ifndef CHANGEOVERTIMESTEPPOSTPROCESSOR_H
 #define CHANGEOVERTIMESTEPPOSTPROCESSOR_H
 
-#include "GeneralPostprocessor.h"
+#include "ChangeOverTimePostprocessor.h"
 
 class ChangeOverTimestepPostprocessor;
 
@@ -24,35 +24,12 @@ InputParameters validParams<ChangeOverTimestepPostprocessor>();
 
 /**
  * Computes the change in a post-processor value, or the magnitude of its
- * relative change, over a time step.
+ * relative change, over a time step or over the entire transient.
  */
-class ChangeOverTimestepPostprocessor : public GeneralPostprocessor
+class ChangeOverTimestepPostprocessor : public ChangeOverTimePostprocessor
 {
 public:
   ChangeOverTimestepPostprocessor(const InputParameters & parameters);
-
-  virtual void initialize() override;
-  virtual void execute() override;
-  virtual Real getValue() override;
-
-protected:
-  /// option to compute change with respect to initial value instead of previous time value
-  const bool _change_with_respect_to_initial;
-
-  /// option to compute the magnitude of relative change instead of change
-  const bool _compute_relative_change;
-
-  /// option to take the absolute value of the change
-  const bool _take_absolute_value;
-
-  /// current post-processor value
-  const PostprocessorValue & _pps_value;
-
-  /// old post-processor value
-  const PostprocessorValue & _pps_value_old;
-
-  /// initial post-processor value
-  Real _pps_value_initial;
 };
 
 #endif /* CHANGEOVERTIMESTEPPOSTPROCESSOR_H */
