+++ test/tests/postprocessors/change_over_time/gold/change_over_time_previous.csv
@@ -1,0 +1,5 @@
+time,change_over_time,my_postprocessor
+0,0,1
+1,1,2
+2,3,5
+
