--- test/tests/postprocessors/change_over_timestep/gold/change_over_timestep_initial.csv
@@ -1,4 +1,0 @@
-time,change_over_timestep,my_postprocessor
-0,0,1
-1,1,2
-2,4,5
