--- test/tests/postprocessors/change_over_timestep/gold/change_over_timestep_previous.csv
@@ -1,5 +1,0 @@
-time,change_over_timestep,my_postprocessor
-0,0,1
-1,1,2
-2,3,5
-
