--- test/tests/postprocessors/change_over_timestep/tests
@@ -1,25 +1,0 @@
-[Tests]
-  [./change_with_respect_to_initial_error_dependent]
-    type = 'RunException'
-    input = 'change_over_timestep.i'
-    cli_args = "Postprocessors/change_over_timestep/change_with_respect_to_initial=true Postprocessors/my_postprocessor/execute_on='timestep_end'"
-    expect_err = "When 'change_with_respect_to_initial' is specified to be true, 'execute_on' for the dependent post-processor \('my_postprocessor'\) must include 'initial'"
-  [../]
-  [./change_with_respect_to_initial_error_this]
-    type = 'RunException'
-    input = 'change_over_timestep.i'
-    cli_args = "Postprocessors/change_over_timestep/change_with_respect_to_initial=true Postprocessors/change_over_timestep/execute_on='timestep_end'"
-    expect_err = "When 'change_with_respect_to_initial' is specified to be true, 'execute_on' for the ChangeOverTimestepPostprocessor \('change_over_timestep'\) must include 'initial'"
-  [../]
-  [./change_with_respect_to_previous]
-    type = 'CSVDiff'
-    input = 'change_over_timestep.i'
-    csvdiff = 'change_over_timestep_previous.csv'
-  [../]
-  [./change_with_respect_to_initial]
-    type = 'CSVDiff'
-    input = 'change_over_timestep.i'
-    cli_args = "Postprocessors/change_over_timestep/change_with_respect_to_initial=true Outputs/file_base='change_over_timestep_initial'"
-    csvdiff = 'change_over_timestep_initial.csv'
-  [../]
-[]
